Increased the maximum fog density to 64

+++ RealisticWeather/RealisticWeatherManager.cs
@@ -3,28 +3,28 @@ using TaleWorlds.Library;
 
 namespace RealisticWeather
 {
-    public class RealisticWeatherManager
-    {
-        private static readonly RealisticWeatherManager _realisticWeatherManager = new RealisticWeatherManager();
+	public class RealisticWeatherManager
+	{
+		private static readonly RealisticWeatherManager _realisticWeatherManager = new RealisticWeatherManager();
 
-        public static RealisticWeatherManager Current => _realisticWeatherManager;
+		public static RealisticWeatherManager Current => _realisticWeatherManager;
 
-        public float RainDensity { get; set; }
+		public float RainDensity { get; set; }
 
-        public float FogDensity { get; set; }
+		public int FogDensity { get; set; }
 
-        public bool HasDust { get; set; }
+		public bool HasDust { get; set; }
 
-        public List<Vec3> WeatherEventPositions { get; set; }
+		public List<Vec3> WeatherEventPositions { get; set; }
 
-        public RealisticWeatherManager() => WeatherEventPositions = new List<Vec3>();
+		public RealisticWeatherManager() => WeatherEventPositions = new List<Vec3>();
 
-        public void SetRainDensity(float rainDensity) => RainDensity = rainDensity;
+		public void SetRainDensity(float rainDensity) => RainDensity = rainDensity;
 
-        public void SetFogDensity(float fogDensity) => FogDensity = fogDensity;
+		public void SetFogDensity(int fogDensity) => FogDensity = fogDensity;
 
-        public void SetDust(bool hasDust) => HasDust = hasDust;
+		public void SetDust(bool hasDust) => HasDust = hasDust;
 
-        public void SetWeatherEventPositions(List<Vec3> weatherEventPositions) => WeatherEventPositions = weatherEventPositions;
-    }
+		public void SetWeatherEventPositions(List<Vec3> weatherEventPositions) => WeatherEventPositions = weatherEventPositions;
+	}
 }
